fix(web): gate /dev/* pages behind a production notFound() guard
All 7 dev-only pages (badge grid/rows/og screenshots, hero-bg, preview,
social) shipped to production reachable by anyone with the URL, with no
guard beyond robots.txt's crawl disallow. Add a NODE_ENV check that calls
notFound() outside development.

preview/social (the two pulling html-to-image into their client bundle)
needed restructuring: their bodies moved to preview-client.tsx/
social-client.tsx, with page.tsx now a plain server component that gates
before rendering the client component — so the guard runs before any
client JS ships, not after it's already loaded in the browser.

Verified live: all 7 routes render "Page not found" content in
production, and preview/social's HTML contains no html-to-image/toPng
reference. Found and logged as a new backlog item (P18) a broader,
pre-existing bug unrelated to this change: notFound() serves HTTP 200
instead of 404 site-wide under `next start`, reproducible on a genuinely
unmatched route with zero backlog changes applied.
