+++ packages/web/app/dev/badge-rows-light/page.tsx
@@ -1,6 +1,7 @@
 // shieldcn — app/dev/badge-rows-light/page.tsx
 // Light mode version of badge rows for OG backgrounds
 
+import { notFound } from "next/navigation"
 import { allBadgePaths } from "@/lib/showcase-data"
 
 function shuffleWithSeed(arr: string[], seed: number): string[] {
@@ -16,6 +17,8 @@ const ROW_COUNT = 9
 const BADGES_PER_ROW = 24
 
 export default function BadgeRowsLightPage() {
+  if (process.env.NODE_ENV !== "development") notFound()
+
   const rows: string[][] = []
   for (let i = 0; i < ROW_COUNT; i++) {
     const shuffled = shuffleWithSeed(allBadgePaths, i * 7 + 3)
