+++ packages/web/app/dev/badge-rows/page.tsx
@@ -2,6 +2,7 @@
 // Horizontal scrolling rows of badges for OG backgrounds
 // Alternating directions like the marquee hero
 
+import { notFound } from "next/navigation"
 import { allBadgePaths } from "@/lib/showcase-data"
 
 function shuffleWithSeed(arr: string[], seed: number): string[] {
@@ -17,6 +18,8 @@ const ROW_COUNT = 9
 const BADGES_PER_ROW = 24
 
 export default function BadgeRowsPage() {
+  if (process.env.NODE_ENV !== "development") notFound()
+
   const rows: string[][] = []
   for (let i = 0; i < ROW_COUNT; i++) {
     const shuffled = shuffleWithSeed(allBadgePaths, i * 7 + 3)
